packages/planum/src/tooltip/tooltip.tsx
import { createContext } from 'react'

import type { TooltipOptions } from './use-tooltip'
import { useTooltip } from './use-tooltip'

type ContextType = ReturnType<typeof useTooltip> | null

export const TooltipContext = createContext<ContextType>(null)

export function Tooltip({
  children,
  ...options
}: { children: React.ReactNode } & TooltipOptions) {
  const tooltip = useTooltip(options)
  return (
    <TooltipContext.Provider value={tooltip}>
      {children}
    </TooltipContext.Provider>
  )
}
